+++ java/com/turbovnc/rfb/AliasParameter.java
@@ -1,4 +1,4 @@
-/* Copyright (C) 2012, 2018, 2022 D. R. Commander.  All Rights Reserved.
+/* Copyright (C) 2012, 2018, 2022-2023 D. R. Commander.  All Rights Reserved.
  * Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
@@ -30,6 +30,8 @@ public final class AliasParameter extends VoidParameter {
   public boolean set(String str) { return param.set(str); }
 
   public void reset() { param.reset(); }
+
+  public boolean setDefault(String str) { return param.setDefault(str); }
 
   public String getDefaultStr() { return null; }
   public String getStr() { return null; }